training

+++ harmonization/swin_contrastive/train.py
@@ -783,7 +783,7 @@ class LazyPatchLoader(Transform):
                 if patch.ndim == 3:
                     patch = patch[np.newaxis, ...]
                 
-                patches.append(patch.double())
+                patches.append(patch)
                 self.logger.info(f"Patch shape: {patch.shape}")
 
             patches = np.concatenate(patches, axis=0)
